backend/requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0
aiosqlite>=0.19.0

# AI - Gemini
google-generativeai>=0.3.0

# Utils
python-dotenv>=1.0.0
python-multipart>=0.0.6
httpx>=0.26.0

# Security & Rate Limiting
slowapi>=0.1.9

# Email
resend>=0.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
